feat: 协议和中英文切换挪位置 # Reviewed, transaction id: 47006
